requirements.txt
python-dotenv
easyocr
requests
PyPDF2
pdf2image
Pillow
openai